requirements.txt
flask
python-dotenv
pandas
scikit-learn
nltk
gunicorn
waitress
